requirements.txt
pytrends>=4.9.2
yfinance>=0.2.40
pandas>=2.2.0
scipy>=1.13.0
statsmodels>=0.14.2
plotly>=5.22.0
pyyaml>=6.0.1
papermill>=2.6.0
